pyproject.toml
[project]
name = "hw1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "certifi>=2026.2.25",
    "google-generativeai>=0.8.6",
    "gradio>=6.8.0",
    "openai>=2.24.0",
    "python-dotenv>=1.2.2",
]
